reorganize dashboard folder and rename file
