+++ src/App.scss
@@ -1,9 +1,3 @@
-.pie-chart {
-  height: 20vh;
-  width: 40vw;
-  margin: 0;
-}
-
 .App-header {
   min-height: 100vh;
   display: flex;
