+++ src/Components/Dashboard/dashboard.scss
@@ -1,7 +1,8 @@
 .dashboard-container {
-padding-left: 150px;
-padding-top: 300px;
+  padding-left: 150px;
+  padding-top: 300px;
 }
+
 .pie-chart {
   position: absolute;
   top: 0;
@@ -11,9 +12,9 @@ padding-top: 300px;
 }
 
 .username-display {
-position: absolute;
-top: 0;
-right: 0;
-padding-right: 20px;
-padding-top:50px;
+  position: absolute;
+  top: 0;
+  right: 0;
+  padding-right: 20px;
+  padding-top:50px;
 }
